r3587: charge is back to integer type

+++ libs/ptc/src/ptcinterface.f90
@@ -139,14 +139,14 @@ CAVITY_TOTALPATH=0
 
         USE orbit_ptc
         IMPLICIT NONE
-        REAL(DP) mass,kin_energy
-        real(dp) charge
-!         integer charge
+        REAL(DP) mass, kin_energy
+!        real(dp) charge
+        integer charge
 
         call GET_MASS_AMU(mass)
         
         call GET_kinetic(kin_energy)
-        call  GET_CHARGE(charge)
+        call GET_CHARGE(charge)
 
       return
       end
